Moving and renaming surface_attributes
New location: within the raster_processing module
New name: surface_attributes renamed to derivatives

call by ...raster.processing.derivatives.function

- Functions checked
- Tests checked
- Notebook checked
- Precommit checked
