+++ eis_toolkit/raster_processing/derivatives/utilities.py
